# internal/server/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Anti Bruteforce",
        "title": "Anti Bruteforce API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "host": "api.anti-bruteforce.local",
    "basePath": "/",
    "paths": {
        "/add/black-list": {
            "post": {
                "description": "Добавить сеть в черный список",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Add"
                ],
                "summary": "Добавить сеть в черный список",
                "parameters": [
                    {
                        "description": "Добавить сеть в черный список",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestIP"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    }
                }
            }
        },
        "/add/white-list": {
            "post": {
                "description": "Добавить сеть в белый список",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Add"
                ],
                "summary": "Добавить сеть в белый список",
                "parameters": [
                    {
                        "description": "Добавить сеть в белый список",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestIP"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    }
                }
            }
        },
        "/check/auth": {
            "post": {
                "description": "Попытка авторизации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Check"
                ],
                "summary": "Попытка авторизации",
                "parameters": [
                    {
                        "description": "Создание события",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestCheck"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/JSONResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    }
                }
            }
        },
        "/delete/black-list": {
            "post": {
                "description": "Удалить сеть из черного списка",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delete"
                ],
                "summary": "Удалить сеть из черного списка",
                "parameters": [
                    {
                        "description": "Удалить сеть из черного списка",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestIP"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    }
                }
            }
        },
        "/delete/bucket": {
            "post": {
                "description": "Удалить бакет",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delete"
                ],
                "summary": "Удалить бакет",
                "parameters": [
                    {
                        "description": "Удалить бакет",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestDeleteBucket"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    }
                }
            }
        },
        "/delete/white-list": {
            "post": {
                "description": "Удалить сеть из белого списка",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delete"
                ],
                "summary": "Удалить сеть из белого списка",
                "parameters": [
                    {
                        "description": "Удалить сеть из белого списка",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestIP"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Пинг-Понг",
                "tags": [
                    "Test"
                ],
                "summary": "Пинг-Понг",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "JSONError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "JSONResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "RequestCheck": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string",
                    "example": "127.0.0.1"
                },
                "login": {
                    "type": "string",
                    "example": "login"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                }
            }
        },
        "RequestDeleteBucket": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string",
                    "example": "127.0.0.1"
                },
                "login": {
                    "type": "string",
                    "example": "login"
                }
            }
        },
        "RequestIP": {
            "type": "object",
            "properties": {
                "net": {
                    "type": "string",
                    "example": "127.0.0.0/24"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}